requirements.txt
asyncclick==8.1.7.2
click==8.1.7
openai==1.51.0
pandas==2.2.3
python-dotenv==1.0.1
python-slugify==8.0.4
unsloth==2024.10.7
pyarrow==17.0.0
joblib==1.4.2
playwright==1.47.0
ollama==0.3.3
